[♻️refactor]: 함수 부분을 커스텀 훅으로 분리
